Fix #20.
When a gcrypt remote is not available, for any reason
(network being down, the drive containing it not being
mounted, etc), pushing to the remote causes a new repository
ID to be created.

Instead, we now fail loudly when a remote is unavailable and
when the remote ID is already set.

https://github.com/bluss/git-remote-gcrypt/issues/20